wallet should no longer deal with account passwords
