+++ validator/accounts/v2/accounts_import.go
@@ -11,12 +11,16 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/k0kubun/go-ansi"
 	"github.com/logrusorgru/aurora"
 	"github.com/pkg/errors"
+	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	v2keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v2"
 	"github.com/prysmaticlabs/prysm/validator/keymanager/v2/direct"
+	"github.com/schollz/progressbar/v3"
 	"github.com/urfave/cli/v2"
+	keystorev4 "github.com/wealdtech/go-eth2-wallet-encryptor-keystorev4"
 )
 
 var derivationPathRegex = regexp.MustCompile("m_12381_3600_([0-9]+)_([0-9]+)_([0-9]+)")
@@ -91,16 +95,13 @@ func ImportAccount(cliCtx *cli.Context) error {
 	if err != nil {
 		return err
 	}
-	km, err := direct.NewKeymanager(ctx, wallet, directCfg)
+	km, err := direct.NewKeymanager(cliCtx, wallet, directCfg)
 	if err != nil {
 		return err
 	}
 	keysDir, err := inputDirectory(cliCtx, importKeysDirPromptText, flags.KeysDirFlag)
 	if err != nil {
 		return errors.Wrap(err, "could not parse keys directory")
-	}
-	if err := wallet.SaveWallet(); err != nil {
-		return errors.Wrap(err, "could not save wallet")
 	}
 	isDir, err := hasDir(keysDir)
 	if err != nil {
@@ -166,6 +167,118 @@ func (w *Wallet) readKeystoreFile(ctx context.Context, keystoreFilePath string) 
 		return nil, errors.Wrap(err, "could not decode keystore json")
 	}
 	return keystoreFile, nil
+}
+
+func (w *Wallet) enterPasswordForAllAccounts(cliCtx *cli.Context, accountNames []string, pubKeys [][]byte) error {
+	au := aurora.NewAurora(true)
+	var password string
+	var err error
+	if cliCtx.IsSet(flags.AccountPasswordFileFlag.Name) {
+		passwordFilePath := cliCtx.String(flags.AccountPasswordFileFlag.Name)
+		data, err := ioutil.ReadFile(passwordFilePath)
+		if err != nil {
+			return err
+		}
+		password = string(data)
+		for i := 0; i < len(accountNames); i++ {
+			err = w.checkPasswordForAccount(accountNames[i], password)
+			if err != nil && strings.Contains(err.Error(), "invalid checksum") {
+				return fmt.Errorf("invalid password for account with public key %#x", pubKeys[i])
+			}
+			if err != nil {
+				return err
+			}
+		}
+	} else {
+		password, err = inputWeakPassword(
+			cliCtx,
+			flags.AccountPasswordFileFlag,
+			"Enter the password for your imported accounts",
+		)
+		fmt.Println("Importing accounts, this may take a while...")
+		bar := progressbar.NewOptions(
+			len(accountNames),
+			progressbar.OptionFullWidth(),
+			progressbar.OptionSetWriter(ansi.NewAnsiStdout()),
+			progressbar.OptionEnableColorCodes(true),
+			progressbar.OptionSetTheme(progressbar.Theme{
+				Saucer:        "[green]=[reset]",
+				SaucerHead:    "[green]>[reset]",
+				SaucerPadding: " ",
+				BarStart:      "[",
+				BarEnd:        "]",
+			}),
+			progressbar.OptionOnCompletion(func() { fmt.Println() }),
+			progressbar.OptionSetDescription("Importing accounts"),
+		)
+		for i := 0; i < len(accountNames); i++ {
+			// We check if the individual account unlocks with the global password.
+			err = w.checkPasswordForAccount(accountNames[i], password)
+			if err != nil && strings.Contains(err.Error(), "invalid checksum") {
+				// If the password fails for an individual account, we ask the user to input
+				// that individual account's password until it succeeds.
+				_, err := w.askUntilPasswordConfirms(cliCtx, accountNames[i], pubKeys[i])
+				if err != nil {
+					return err
+				}
+				if err := bar.Add(1); err != nil {
+					return errors.Wrap(err, "could not add to progress bar")
+				}
+				continue
+			}
+			if err != nil {
+				return err
+			}
+			fmt.Printf("Finished importing %#x\n", au.BrightMagenta(bytesutil.Trunc(pubKeys[i])))
+			if err := bar.Add(1); err != nil {
+				return errors.Wrap(err, "could not add to progress bar")
+			}
+		}
+	}
+	return nil
+}
+
+func (w *Wallet) askUntilPasswordConfirms(cliCtx *cli.Context, accountName string, pubKey []byte) (string, error) {
+	// Loop asking for the password until the user enters it correctly.
+	var password string
+	var err error
+	for {
+		password, err = inputWeakPassword(
+			cliCtx,
+			flags.AccountPasswordFileFlag,
+			fmt.Sprintf(passwordForAccountPromptText, bytesutil.Trunc(pubKey)),
+		)
+		if err != nil {
+			return "", errors.Wrap(err, "could not input password")
+		}
+		err = w.checkPasswordForAccount(accountName, password)
+		if err != nil && strings.Contains(err.Error(), "invalid checksum") {
+			fmt.Println(au.Red("Incorrect password entered, please try again"))
+			continue
+		}
+		if err != nil {
+			return "", err
+		}
+		break
+	}
+	return password, nil
+}
+
+func (w *Wallet) checkPasswordForAccount(accountName string, password string) error {
+	encoded, err := w.ReadFileAtPath(context.Background(), accountName, direct.KeystoreFileName)
+	if err != nil {
+		return errors.Wrap(err, "could not read keystore file")
+	}
+	keystoreJSON := &v2keymanager.Keystore{}
+	if err := json.Unmarshal(encoded, &keystoreJSON); err != nil {
+		return errors.Wrap(err, "could not decode json")
+	}
+	decryptor := keystorev4.New()
+	_, err = decryptor.Decrypt(keystoreJSON.Crypto, password)
+	if err != nil {
+		return errors.Wrap(err, "could not decrypt keystore")
+	}
+	return nil
 }
 
 // Extracts the account index, j, from a derivation path in a file name
