+++ validator/accounts/v2/iface/wallet.go
@@ -12,7 +12,6 @@ type Wallet interface {
 	// Methods to retrieve wallet and accounts metadata.
 	AccountsDir() string
 	ListDirs() ([]string, error)
-	Password() string
 	// Read methods for important wallet and accounts-related files.
 	ReadEncryptedSeedFromDisk(ctx context.Context) (io.ReadCloser, error)
 	ReadFileAtPath(ctx context.Context, filePath string, fileName string) ([]byte, error)
