+++ validator/accounts/v2/wallet.go
@@ -2,31 +2,21 @@ package v2
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strconv"
-	"strings"
-	"time"
 
 	petname "github.com/dustinkirkland/golang-petname"
-	"github.com/k0kubun/go-ansi"
-	"github.com/logrusorgru/aurora"
 	"github.com/pkg/errors"
-	"github.com/prysmaticlabs/prysm/shared/bytesutil"
-	"github.com/prysmaticlabs/prysm/shared/promptutil"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	v2keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v2"
 	"github.com/prysmaticlabs/prysm/validator/keymanager/v2/derived"
 	"github.com/prysmaticlabs/prysm/validator/keymanager/v2/direct"
 	"github.com/prysmaticlabs/prysm/validator/keymanager/v2/remote"
-	"github.com/schollz/progressbar/v3"
 	"github.com/sirupsen/logrus"
 	"github.com/urfave/cli/v2"
-	keystorev4 "github.com/wealdtech/go-eth2-wallet-encryptor-keystorev4"
 )
 
 const (
@@ -91,25 +81,11 @@ func NewWallet(
 		}
 	}
 	accountsPath := filepath.Join(walletDir, keymanagerKind.String())
-	w := &Wallet{
+	return &Wallet{
 		accountsPath:   accountsPath,
 		keymanagerKind: keymanagerKind,
 		walletDir:      walletDir,
-	}
-	if keymanagerKind == v2keymanager.Derived || keymanagerKind == v2keymanager.Direct {
-		walletPassword, err := inputPassword(
-			cliCtx,
-			flags.WalletPasswordFileFlag,
-			newWalletPasswordPromptText,
-			confirmPass,
-			promptutil.ValidatePasswordInput,
-		)
-		if err != nil {
-			return nil, errors.Wrap(err, "could not get password")
-		}
-		w.walletPassword = walletPassword
-	}
-	return w, nil
+	}, nil
 }
 
 // OpenWallet instantiates a wallet from a specified path. It checks the
@@ -141,71 +117,12 @@ func OpenWallet(cliCtx *cli.Context) (*Wallet, error) {
 		return nil, errors.Wrap(err, "could not read keymanager kind for wallet")
 	}
 	walletPath := filepath.Join(walletDir, keymanagerKind.String())
-	w := &Wallet{
+	log.Infof("%s %s", au.BrightMagenta("(wallet directory)"), walletDir)
+	return &Wallet{
 		walletDir:      walletDir,
 		accountsPath:   walletPath,
 		keymanagerKind: keymanagerKind,
-	}
-	// Check if the wallet is using the new, fast keystore format.
-	hasNewFormat, err := hasDir(filepath.Join(walletPath, direct.AccountsPath))
-	if err != nil {
-		return nil, errors.Wrap(err, "could not read wallet dir")
-	}
-	log.Infof("%s %s", au.BrightMagenta("(wallet directory)"), w.walletDir)
-	if keymanagerKind == v2keymanager.Derived {
-		validateExistingPass := func(input string) error {
-			if input == "" {
-				return errors.New("password input cannot be empty")
-			}
-			return nil
-		}
-		walletPassword, err := inputPassword(
-			cliCtx,
-			flags.WalletPasswordFileFlag,
-			walletPasswordPromptText,
-			noConfirmPass,
-			validateExistingPass,
-		)
-		if err != nil {
-			return nil, err
-		}
-		w.walletPassword = walletPassword
-	}
-	if keymanagerKind == v2keymanager.Direct {
-		var walletPassword string
-		if hasNewFormat {
-			validateExistingPass := func(input string) error {
-				if input == "" {
-					return errors.New("password input cannot be empty")
-				}
-				return nil
-			}
-			walletPassword, err = inputPassword(
-				cliCtx,
-				flags.WalletPasswordFileFlag,
-				walletPasswordPromptText,
-				noConfirmPass,
-				validateExistingPass,
-			)
-		} else {
-			fmt.Println("\nWe have revamped how imported accounts work, improving speed significantly for your " +
-				"validators as well as reducing memory and CPU requirements. This unifies all your existing accounts " +
-				"into a single format protected by a strong password. You'll need to set a new password for this " +
-				"updated wallet format")
-			walletPassword, err = inputPassword(
-				cliCtx,
-				flags.WalletPasswordFileFlag,
-				newWalletPasswordPromptText,
-				confirmPass,
-				promptutil.ValidatePasswordInput,
-			)
-		}
-		if err != nil {
-			return nil, err
-		}
-		w.walletPassword = walletPassword
-	}
-	return w, nil
+	}, nil
 }
 
 // SaveWallet persists the wallet's directories to disk.
@@ -226,17 +143,13 @@ func (w *Wallet) AccountsDir() string {
 	return w.accountsPath
 }
 
-// Password for the wallet.
-func (w *Wallet) Password() string {
-	return w.walletPassword
-}
-
 // InitializeKeymanager reads a keymanager config from disk at the wallet path,
 // unmarshals it based on the wallet's keymanager kind, and returns its value.
 func (w *Wallet) InitializeKeymanager(
-	ctx context.Context,
+	cliCtx *cli.Context,
 	skipMnemonicConfirm bool,
 ) (v2keymanager.IKeymanager, error) {
+	ctx := context.Background()
 	configFile, err := w.ReadKeymanagerConfigFromDisk(ctx)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not read keymanager config")
@@ -248,7 +161,7 @@ func (w *Wallet) InitializeKeymanager(
 		if err != nil {
 			return nil, errors.Wrap(err, "could not unmarshal keymanager config file")
 		}
-		keymanager, err = direct.NewKeymanager(ctx, w, cfg)
+		keymanager, err = direct.NewKeymanager(cliCtx, w, cfg)
 		if err != nil {
 			return nil, errors.Wrap(err, "could not initialize direct keymanager")
 		}
@@ -257,7 +170,7 @@ func (w *Wallet) InitializeKeymanager(
 		if err != nil {
 			return nil, errors.Wrap(err, "could not unmarshal keymanager config file")
 		}
-		keymanager, err = derived.NewKeymanager(ctx, w, cfg, skipMnemonicConfirm, w.walletPassword)
+		keymanager, err = derived.NewKeymanager(cliCtx, w, cfg, skipMnemonicConfirm, w.walletPassword)
 		if err != nil {
 			return nil, errors.Wrap(err, "could not initialize derived keymanager")
 		}
@@ -266,7 +179,7 @@ func (w *Wallet) InitializeKeymanager(
 		if err != nil {
 			return nil, errors.Wrap(err, "could not unmarshal keymanager config file")
 		}
-		keymanager, err = remote.NewKeymanager(ctx, 100000000, cfg)
+		keymanager, err = remote.NewKeymanager(cliCtx, 100000000, cfg)
 		if err != nil {
 			return nil, errors.Wrap(err, "could not initialize remote keymanager")
 		}
@@ -364,22 +277,6 @@ func (w *Wallet) FileNameAtPath(ctx context.Context, filePath string, fileName s
 	return fullFileName, nil
 }
 
-// AccountTimestamp retrieves the timestamp from a given keystore file name.
-func AccountTimestamp(fileName string) (time.Time, error) {
-	timestampStart := strings.LastIndex(fileName, "-") + 1
-	timestampEnd := strings.LastIndex(fileName, ".")
-	// Return an error if the text we expect cannot be found.
-	if timestampStart == -1 || timestampEnd == -1 {
-		return time.Unix(0, 0), fmt.Errorf("could not find timestamp in file name %s", fileName)
-	}
-	unixTimestampStr, err := strconv.ParseInt(fileName[timestampStart:timestampEnd], 10, 64)
-	if err != nil {
-		return time.Unix(0, 0), errors.Wrapf(err, "could not parse account created at timestamp: %s", fileName)
-	}
-	unixTimestamp := time.Unix(unixTimestampStr, 0)
-	return unixTimestamp, nil
-}
-
 // ReadKeymanagerConfigFromDisk opens a keymanager config file
 // for reading if it exists at the wallet path.
 func (w *Wallet) ReadKeymanagerConfigFromDisk(ctx context.Context) (io.ReadCloser, error) {
@@ -421,168 +318,6 @@ func (w *Wallet) WriteEncryptedSeedToDisk(ctx context.Context, encoded []byte) e
 		return errors.Wrapf(err, "could not write %s", seedFilePath)
 	}
 	log.WithField("seedFilePath", seedFilePath).Debug("Wrote wallet encrypted seed file to disk")
-	return nil
-}
-
-// enterPasswordForAccount checks if a user has a password specified for the new account
-// either from a file or from stdin. Then, it saves the password to the wallet.
-func (w *Wallet) enterPasswordForAccount(cliCtx *cli.Context, accountName string, pubKey []byte) error {
-	au := aurora.NewAurora(true)
-	var password string
-	var err error
-	if cliCtx.IsSet(flags.AccountPasswordFileFlag.Name) {
-		passwordFilePath := cliCtx.String(flags.AccountPasswordFileFlag.Name)
-		data, err := ioutil.ReadFile(passwordFilePath)
-		if err != nil {
-			return err
-		}
-		password = string(data)
-		err = w.checkPasswordForAccount(accountName, password)
-		if err != nil && strings.Contains(err.Error(), "invalid checksum") {
-			return fmt.Errorf("invalid password entered for account with public key %#x", pubKey)
-		}
-		if err != nil {
-			return err
-		}
-	} else {
-		pubKeyStr := fmt.Sprintf("%#x", bytesutil.Trunc(pubKey))
-		attemptingPassword := true
-		// Loop asking for the password until the user enters it correctly.
-		for attemptingPassword {
-			// Ask the user for the password to their account.
-			password, err = inputWeakPassword(
-				cliCtx,
-				flags.AccountPasswordFileFlag,
-				fmt.Sprintf(passwordForAccountPromptText, au.BrightGreen(pubKeyStr)),
-			)
-			if err != nil {
-				return errors.Wrap(err, "could not input password")
-			}
-			err = w.checkPasswordForAccount(accountName, password)
-			if err != nil && strings.Contains(err.Error(), "invalid checksum") {
-				fmt.Print(au.Red("X").Bold())
-				fmt.Print(au.Red("\nIncorrect password entered, please try again"))
-				continue
-			}
-			if err != nil {
-				return err
-			}
-			attemptingPassword = false
-			fmt.Print(au.Green("✔️\n").Bold())
-		}
-	}
-	return nil
-}
-
-func (w *Wallet) enterPasswordForAllAccounts(cliCtx *cli.Context, accountNames []string, pubKeys [][]byte) error {
-	au := aurora.NewAurora(true)
-	var password string
-	var err error
-	if cliCtx.IsSet(flags.AccountPasswordFileFlag.Name) {
-		passwordFilePath := cliCtx.String(flags.AccountPasswordFileFlag.Name)
-		data, err := ioutil.ReadFile(passwordFilePath)
-		if err != nil {
-			return err
-		}
-		password = string(data)
-		for i := 0; i < len(accountNames); i++ {
-			err = w.checkPasswordForAccount(accountNames[i], password)
-			if err != nil && strings.Contains(err.Error(), "invalid checksum") {
-				return fmt.Errorf("invalid password for account with public key %#x", pubKeys[i])
-			}
-			if err != nil {
-				return err
-			}
-		}
-	} else {
-		password, err = inputWeakPassword(
-			cliCtx,
-			flags.AccountPasswordFileFlag,
-			"Enter the password for your imported accounts",
-		)
-		fmt.Println("Importing accounts, this may take a while...")
-		bar := progressbar.NewOptions(
-			len(accountNames),
-			progressbar.OptionFullWidth(),
-			progressbar.OptionSetWriter(ansi.NewAnsiStdout()),
-			progressbar.OptionEnableColorCodes(true),
-			progressbar.OptionSetTheme(progressbar.Theme{
-				Saucer:        "[green]=[reset]",
-				SaucerHead:    "[green]>[reset]",
-				SaucerPadding: " ",
-				BarStart:      "[",
-				BarEnd:        "]",
-			}),
-			progressbar.OptionOnCompletion(func() { fmt.Println() }),
-			progressbar.OptionSetDescription("Importing accounts"),
-		)
-		for i := 0; i < len(accountNames); i++ {
-			// We check if the individual account unlocks with the global password.
-			err = w.checkPasswordForAccount(accountNames[i], password)
-			if err != nil && strings.Contains(err.Error(), "invalid checksum") {
-				// If the password fails for an individual account, we ask the user to input
-				// that individual account's password until it succeeds.
-				_, err := w.askUntilPasswordConfirms(cliCtx, accountNames[i], pubKeys[i])
-				if err != nil {
-					return err
-				}
-				if err := bar.Add(1); err != nil {
-					return errors.Wrap(err, "could not add to progress bar")
-				}
-				continue
-			}
-			if err != nil {
-				return err
-			}
-			fmt.Printf("Finished importing %#x\n", au.BrightMagenta(bytesutil.Trunc(pubKeys[i])))
-			if err := bar.Add(1); err != nil {
-				return errors.Wrap(err, "could not add to progress bar")
-			}
-		}
-	}
-	return nil
-}
-
-func (w *Wallet) askUntilPasswordConfirms(cliCtx *cli.Context, accountName string, pubKey []byte) (string, error) {
-	// Loop asking for the password until the user enters it correctly.
-	var password string
-	var err error
-	for {
-		password, err = inputWeakPassword(
-			cliCtx,
-			flags.AccountPasswordFileFlag,
-			fmt.Sprintf(passwordForAccountPromptText, bytesutil.Trunc(pubKey)),
-		)
-		if err != nil {
-			return "", errors.Wrap(err, "could not input password")
-		}
-		err = w.checkPasswordForAccount(accountName, password)
-		if err != nil && strings.Contains(err.Error(), "invalid checksum") {
-			fmt.Println(au.Red("Incorrect password entered, please try again"))
-			continue
-		}
-		if err != nil {
-			return "", err
-		}
-		break
-	}
-	return password, nil
-}
-
-func (w *Wallet) checkPasswordForAccount(accountName string, password string) error {
-	encoded, err := w.ReadFileAtPath(context.Background(), accountName, direct.KeystoreFileName)
-	if err != nil {
-		return errors.Wrap(err, "could not read keystore file")
-	}
-	keystoreJSON := &v2keymanager.Keystore{}
-	if err := json.Unmarshal(encoded, &keystoreJSON); err != nil {
-		return errors.Wrap(err, "could not decode json")
-	}
-	decryptor := keystorev4.New()
-	_, err = decryptor.Decrypt(keystoreJSON.Crypto, password)
-	if err != nil {
-		return errors.Wrap(err, "could not decrypt keystore")
-	}
 	return nil
 }
