+++ validator/accounts/v2/wallet_test.go
@@ -8,10 +8,11 @@ import (
 	"math/big"
 	"os"
 	"path/filepath"
-	"reflect"
 	"strconv"
 	"testing"
-	"time"
+
+	"github.com/sirupsen/logrus"
+	"github.com/urfave/cli/v2"
 
 	"github.com/prysmaticlabs/prysm/shared/params"
 	"github.com/prysmaticlabs/prysm/shared/testutil"
@@ -19,8 +20,6 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/testutil/require"
 	"github.com/prysmaticlabs/prysm/validator/flags"
 	v2keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v2"
-	"github.com/sirupsen/logrus"
-	"github.com/urfave/cli/v2"
 )
 
 func init() {
@@ -86,44 +85,6 @@ func setupWalletAndPasswordsDir(t testing.TB) (string, string, string) {
 		require.NoError(t, os.RemoveAll(passwordsDir), "Failed to remove directory")
 	})
 	return walletDir, passwordsDir, passwordFilePath
-}
-
-func TestAccountTimestamp(t *testing.T) {
-	tests := []struct {
-		name     string
-		fileName string
-		want     time.Time
-		wantErr  bool
-	}{
-		{
-			name:     "keystore with timestamp",
-			fileName: "keystore-1234567.json",
-			want:     time.Unix(1234567, 0),
-		},
-		{
-			name:     "keystore with deriv path and timestamp",
-			fileName: "keystore-12313-313-00-0-5500550.json",
-			want:     time.Unix(5500550, 0),
-		},
-		{
-			name:     "keystore with no timestamp",
-			fileName: "keystore.json",
-			want:     time.Unix(0, 0),
-			wantErr:  true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := AccountTimestamp(tt.fileName)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("AccountTimestamp() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("AccountTimestamp() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
 
 func Test_IsEmptyWallet_RandomFiles(t *testing.T) {
