+++ validator/keymanager/v2/direct/direct.go
@@ -6,6 +6,10 @@ import (
 	"fmt"
 	"io"
 	"io/ioutil"
+	"os"
+	"os/user"
+	"path"
+	"path/filepath"
 	"strings"
 	"sync"
 
@@ -19,8 +23,10 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/petnames"
 	"github.com/prysmaticlabs/prysm/shared/promptutil"
 	"github.com/prysmaticlabs/prysm/validator/accounts/v2/iface"
+	"github.com/prysmaticlabs/prysm/validator/flags"
 	v2keymanager "github.com/prysmaticlabs/prysm/validator/keymanager/v2"
 	"github.com/sirupsen/logrus"
+	"github.com/urfave/cli/v2"
 	keystorev4 "github.com/wealdtech/go-eth2-wallet-encryptor-keystorev4"
 )
 
@@ -31,13 +37,22 @@ const (
 	KeystoreFileName = "keystore-*.json"
 	// KeystoreFileNameFormat exposes the filename the keystore should be formatted in.
 	KeystoreFileNameFormat = "keystore-%d.json"
-	// PasswordFileSuffix for passwords persisted as text to disk.
-	PasswordFileSuffix = ".pass"
 	// AccountsPath where all direct keymanager keystores are kept.
-	AccountsPath                   = "accounts"
-	accountsKeystoreFileName       = "all-accounts.keystore.json"
-	accountsKeystoreFileNameFormat = "all-accounts.keystore.json"
-	eipVersion                     = "EIP-2335"
+	AccountsPath                = "accounts"
+	accountsKeystoreFileName    = "all-accounts.keystore.json"
+	eipVersion                  = "EIP-2335"
+	newWalletPasswordPromptText = "New wallet password"
+	walletPasswordPromptText    = "Wallet password"
+	confirmPasswordPromptText   = "Confirm password"
+)
+
+type passwordConfirm int
+
+const (
+	// An enum to indicate to the prompt that confirming the password is not needed.
+	noConfirmPass passwordConfirm = iota
+	// An enum to indicate to the prompt to confirm the password entered.
+	confirmPass
 )
 
 // Config for a direct keymanager.
@@ -47,11 +62,12 @@ type Config struct {
 
 // Keymanager implementation for direct keystores utilizing EIP-2335.
 type Keymanager struct {
-	wallet        iface.Wallet
-	cfg           *Config
-	keysCache     map[[48]byte]bls.SecretKey
-	accountsStore *AccountStore
-	lock          sync.RWMutex
+	wallet           iface.Wallet
+	cfg              *Config
+	keysCache        map[[48]byte]bls.SecretKey
+	accountsStore    *AccountStore
+	lock             sync.RWMutex
+	accountsPassword string
 }
 
 // AccountStore --
@@ -68,13 +84,46 @@ func DefaultConfig() *Config {
 }
 
 // NewKeymanager instantiates a new direct keymanager from configuration options.
-func NewKeymanager(ctx context.Context, wallet iface.Wallet, cfg *Config) (*Keymanager, error) {
+func NewKeymanager(ctx *cli.Context, wallet iface.Wallet, cfg *Config) (*Keymanager, error) {
 	k := &Keymanager{
 		wallet:        wallet,
 		cfg:           cfg,
 		keysCache:     make(map[[48]byte]bls.SecretKey),
 		accountsStore: &AccountStore{},
 	}
+
+	walletFiles, err := wallet.ListDirs()
+	if err != nil {
+		return nil, err
+	}
+	var accountsPassword string
+	if len(walletFiles) == 0 {
+		accountsPassword, err = inputPassword(
+			ctx,
+			flags.WalletPasswordFileFlag,
+			newWalletPasswordPromptText,
+			confirmPass,
+			promptutil.ValidatePasswordInput,
+		)
+	} else {
+		validateExistingPass := func(input string) error {
+			if input == "" {
+				return errors.New("password input cannot be empty")
+			}
+			return nil
+		}
+		accountsPassword, err = inputPassword(
+			ctx,
+			flags.WalletPasswordFileFlag,
+			walletPasswordPromptText,
+			noConfirmPass,
+			validateExistingPass,
+		)
+	}
+	if err != nil {
+		return nil, err
+	}
+	k.accountsPassword = accountsPassword
 
 	// If the wallet has the capability of unlocking accounts using
 	// passphrases, then we initialize a cache of public key -> secret keys
@@ -238,8 +287,8 @@ func (dr *Keymanager) Sign(ctx context.Context, req *validatorpb.SignRequest) (b
 	return secretKey.Sign(req.SigningRoot), nil
 }
 
-func (dr *Keymanager) initializeSecretKeysCache(ctx context.Context) error {
-	encoded, err := dr.wallet.ReadFileAtPath(ctx, AccountsPath, accountsKeystoreFileName)
+func (dr *Keymanager) initializeSecretKeysCache(cliCtx *cli.Context) error {
+	encoded, err := dr.wallet.ReadFileAtPath(context.Background(), AccountsPath, accountsKeystoreFileName)
 	if err != nil && strings.Contains(err.Error(), "no files found") {
 		// If there are no keys to initialize at all, just exit.
 		return nil
@@ -254,11 +303,11 @@ func (dr *Keymanager) initializeSecretKeysCache(ctx context.Context) error {
 	// by utilizing the password and initialize a new BLS secret key from
 	// its raw bytes.
 	decryptor := keystorev4.New()
-	enc, err := decryptor.Decrypt(keystoreFile.Crypto, dr.wallet.Password())
+	enc, err := decryptor.Decrypt(keystoreFile.Crypto, dr.accountsPassword)
 	if err != nil && strings.Contains(err.Error(), "invalid checksum") {
 		// If the password fails for an individual account, we ask the user to input
 		// that individual account's password until it succeeds.
-		enc, _, err = dr.askUntilPasswordConfirms(decryptor, keystoreFile)
+		enc, dr.accountsPassword, err = dr.askUntilPasswordConfirms(decryptor, keystoreFile)
 		if err != nil {
 			return errors.Wrap(err, "could not confirm password via prompt")
 		}
@@ -336,7 +385,7 @@ func (dr *Keymanager) createAccountsKeystore(
 	if err != nil {
 		return nil, err
 	}
-	cryptoFields, err := encryptor.Encrypt(encodedStore, dr.wallet.Password())
+	cryptoFields, err := encryptor.Encrypt(encodedStore, dr.accountsPassword)
 	if err != nil {
 		return nil, errors.Wrap(err, "could not encrypt accounts")
 	}
@@ -357,6 +406,7 @@ func (dr *Keymanager) askUntilPasswordConfirms(
 	var password string
 	var err error
 	for {
+		fmt.Println("Asking for password here")
 		password, err = promptutil.PasswordPrompt(
 			"Wrong password entered, try again", promptutil.NotEmpty,
 		)
@@ -374,4 +424,77 @@ func (dr *Keymanager) askUntilPasswordConfirms(
 		break
 	}
 	return secretKey, password, nil
+}
+
+func inputPassword(
+	cliCtx *cli.Context,
+	passwordFileFlag *cli.StringFlag,
+	promptText string,
+	confirmPassword passwordConfirm,
+	passwordValidator func(input string) error,
+) (string, error) {
+	if cliCtx.IsSet(passwordFileFlag.Name) {
+		passwordFilePathInput := cliCtx.String(passwordFileFlag.Name)
+		passwordFilePath, err := expandPath(passwordFilePathInput)
+		if err != nil {
+			return "", errors.Wrap(err, "could not determine absolute path of password file")
+		}
+		data, err := ioutil.ReadFile(passwordFilePath)
+		if err != nil {
+			return "", errors.Wrap(err, "could not read password file")
+		}
+		enteredPassword := strings.TrimRight(string(data), "\r\n")
+		if err := passwordValidator(enteredPassword); err != nil {
+			return "", errors.Wrap(err, "password did not pass validation")
+		}
+		return enteredPassword, nil
+	}
+	var hasValidPassword bool
+	var walletPassword string
+	var err error
+	for !hasValidPassword {
+		walletPassword, err = promptutil.PasswordPrompt(promptText, passwordValidator)
+		if err != nil {
+			return "", fmt.Errorf("could not read account password: %v", err)
+		}
+
+		if confirmPassword == confirmPass {
+			passwordConfirmation, err := promptutil.PasswordPrompt(confirmPasswordPromptText, passwordValidator)
+			if err != nil {
+				return "", fmt.Errorf("could not read password confirmation: %v", err)
+			}
+			if walletPassword != passwordConfirmation {
+				log.Error("Passwords do not match")
+				continue
+			}
+			hasValidPassword = true
+		} else {
+			return walletPassword, nil
+		}
+	}
+	return walletPassword, nil
+}
+
+// Expands a file path
+// 1. replace tilde with users home dir
+// 2. expands embedded environment variables
+// 3. cleans the path, e.g. /a/b/../c -> /a/c
+// Note, it has limitations, e.g. ~someuser/tmp will not be expanded
+func expandPath(p string) (string, error) {
+	if strings.HasPrefix(p, "~/") || strings.HasPrefix(p, "~\\") {
+		if home := homeDir(); home != "" {
+			p = home + p[1:]
+		}
+	}
+	return filepath.Abs(path.Clean(os.ExpandEnv(p)))
+}
+
+func homeDir() string {
+	if home := os.Getenv("HOME"); home != "" {
+		return home
+	}
+	if usr, err := user.Current(); err == nil {
+		return usr.HomeDir
+	}
+	return ""
 }
