+++ validator/node/node.go
@@ -88,7 +88,7 @@ func NewValidatorClient(cliCtx *cli.Context) (*ValidatorClient, error) {
 			log.Fatalf("Could not open wallet: %v", err)
 		}
 		keyManagerV2, err = wallet.InitializeKeymanager(
-			context.Background(), false, /* skipMnemonicConfirm */
+			cliCtx, false, /* skipMnemonicConfirm */
 		)
 		if err != nil {
 			log.Fatalf("Could not read existing keymanager for wallet: %v", err)
